requirements.txt
fastapi
uvicorn
scikit-learn
python-dotenv
openai
anthropic